+++ example/github_trend_example.dart
@@ -1,36 +1,36 @@
-import 'package:github_trend/github_trending.dart';
+import 'package:github_trend/github_trend.dart';
 import 'dart:convert';
 
 main() async {
-  GithubTrending githubTrending = GithubTrending();
+  GithubTrend githubTrend = GithubTrend();
 
   // fetch languages
-  List<String> languages = await githubTrending.fetchLanguages();
+  List<String> languages = await githubTrend.fetchLanguages();
   print(languages);
   print(json.encode(languages));
   // default time language
   try {
     List<Map<String, dynamic>> defaultArray =
-        await githubTrending.fetchTrendingRepos();
+        await githubTrend.fetchTrendingRepos();
     print(defaultArray);
   } catch (e) {
     // get origin response
-    githubTrending.response;
+    githubTrend.response;
     print(e);
   }
 
   // rust language and default time
   List<Map<String, dynamic>> rustArray =
-      await githubTrending.fetchTrendingRepos(language: 'rust');
+      await githubTrend.fetchTrendingRepos(language: 'rust');
   print(rustArray);
 
   // rust language and last weekly
-  List<Map<String, dynamic>> rustWeekArray = await githubTrending
+  List<Map<String, dynamic>> rustWeekArray = await githubTrend
       .fetchTrendingRepos(language: 'rust', since: 'weekly');
   print(rustWeekArray);
 
   // return default trending when pass error time and language params.
-  List<Map<String, dynamic>> rustWeekArrayerror = await githubTrending
+  List<Map<String, dynamic>> rustWeekArrayerror = await githubTrend
       .fetchTrendingRepos(language: 'error', since: 'error');
   print(rustWeekArrayerror);
 }