+++ lib/github_trend.dart
@@ -1,8 +1,8 @@
 /// Support for doing something awesome.
 ///
 /// More dartdocs go here.
-library github_trending;
+library github_trend;
 
-export 'src/github_trending_base.dart';
+export 'src/github_trend_base.dart';
 
 // TODO: Export any libraries intended for clients of this package.